+++ src/data.ts
@@ -1,7 +1,7 @@
 /**
  *	MIT License
  *
- *	Copyright (c) 2019 - 2021 Toreda, Inc.
+ *	Copyright (c) 2019 - 2022 Toreda, Inc.
  *
  *	Permission is hereby granted, free of charge, to any person obtaining a copy
  *	of this software and associated documentation files (the "Software"), to deal
@@ -24,8 +24,8 @@
  */
 
 /**
- * Expressive type alias for basic Toreda data object types.
+ * Describes dictionary objects mapping strings to primitive data.
  *
  * @category Expressive Types
  */
-export type Data = Record<string, unknown>;
+export type Data = Record<string, string | boolean | null | number | unknown[]>;