+++ src/deep/expand.ts
@@ -1,7 +1,7 @@
 /**
  *	MIT License
  *
- *	Copyright (c) 2019 - 2021 Toreda, Inc.
+ *	Copyright (c) 2019 - 2022 Toreda, Inc.
  *
  *	Permission is hereby granted, free of charge, to any person obtaining a copy
  *	of this software and associated documentation files (the "Software"), to deal
@@ -26,7 +26,7 @@
 import {Primitive} from '../primitive';
 
 /**
- * @category Deep
+ * @category Object APIs
  */
 export type DeepExpand<T> = T extends Primitive
 	? T