+++ src/deep/required.ts
@@ -1,7 +1,7 @@
 /**
  *	MIT License
  *
- *	Copyright (c) 2019 - 2021 Toreda, Inc.
+ *	Copyright (c) 2019 - 2022 Toreda, Inc.
  *
  *	Permission is hereby granted, free of charge, to any person obtaining a copy
  *	of this software and associated documentation files (the "Software"), to deal
@@ -27,7 +27,7 @@ import {Expand} from '../expand';
 import {Primitive} from '../primitive';
 
 /**
- * @category Deep
+ * @category Object APIs
  */
 export type DeepRequired<T> = Expand<{
 	[key in keyof T]-?: T[key] extends Primitive ? NonNullable<T[key]> : DeepRequired<T[key]>;