+++ src/expand.ts
@@ -1,7 +1,7 @@
 /**
  *	MIT License
  *
- *	Copyright (c) 2019 - 2021 Toreda, Inc.
+ *	Copyright (c) 2019 - 2022 Toreda, Inc.
  *
  *	Permission is hereby granted, free of charge, to any person obtaining a copy
  *	of this software and associated documentation files (the "Software"), to deal
@@ -24,6 +24,6 @@
  */
 
 /**
- * @category
+ * @category Expressive Types
  */
 export type Expand<T> = T extends infer obj ? {[key in keyof obj]: obj[key]} : never;