+++ src/lifecycle/phase/id.ts
@@ -1,0 +1,6 @@
+/**
+ * Used for arguments specifying a target lifecycle phase.
+ *
+ * @category Expressive Types
+ */
+export type LifecyclePhaseId = string;