+++ src/stringable.ts
@@ -1,7 +1,7 @@
 /**
  *	MIT License
  *
- *	Copyright (c) 2019 - 2021 Toreda, Inc.
+ *	Copyright (c) 2019 - 2022 Toreda, Inc.
  *
  *	Permission is hereby granted, free of charge, to any person obtaining a copy
  *	of this software and associated documentation files (the "Software"), to deal
@@ -27,7 +27,7 @@
  * Stringable objects support stringifying its contents and includes a toString
  *  method to do so. Used in conjunction with other types.
  *
- * @category Attribute Types
+ * @category Object APIs
  */
 export interface Stringable {
 	toString: () => string;